backend/pyproject.toml
[project]
name = "faitness"
version = "0.1.0"
description = ""
authors = [
    {name = "megara.arevaco"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn app.main:app --reload"